requirements.txt
# Core libraries for scraping and data processing
pandas==2.2.1
requests==2.31.0
PyYAML==6.0.1
pyarrow==15.0.0

# Library for ticker generation and its sub-dependencies
vnstock==0.2.9.1
beautifulsoup4==4.12.3
packaging==23.2

# Data Visualization & Notebooks
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0